+++ src/App.css
@@ -1,7 +1,7 @@
 @import url('https://fonts.googleapis.com/css2?family=Righteous&family=Roboto+Condensed:wght@300&display=swap');
 
 body {
-  font-family: 'Roboto Condensed', sans-serif;
+  font-family: 'Roboto Condensed', sans-serif !important;
 }
 
 :root {
@@ -84,8 +84,6 @@ nav ul li {
   justify-content: center;
   align-items: center;
   text-align: center;
-  font-family: 'Roboto Condensed', sans-serif;
-
 }
 
 .main-info h1 {
@@ -105,4 +103,30 @@ nav ul li {
   padding: 0.625rem 0.75rem 0.75rem 0.75rem;
   color: var(--primary-white);
   margin: 2rem;
+}
+
+.navbar {
+  z-index: 1;
+}
+
+canvas {
+  height: 100vh !important;
+  position: absolute !important;
+}
+
+.profile-img {
+  border: 1px solid var(--primary-orange);
+  border-radius: 50%;
+  padding: .3rem;
+  width: 16rem;
+}
+
+.about-header {
+  color: var(--primary-orange);
+  text-transform: uppercase;
+}
+.photo-wrap{
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
